refactor(gui): integrate windows app patch modules
Replace runtime monkey-patch imports with a canonical app entrypoint and focused Windows app helper modules.

Update the cleanup plan, README, and unreleased notes for the GUI architecture cleanup.
