+++ run_windows_remote_app.py
@@ -1,5 +1,5 @@
-from windows_app import processing_options_patches as _processing_options_patches
-from windows_app.live_webcam_patches import main
+from windows_app.app import main
+
 
 if __name__ == "__main__":
     raise SystemExit(main())
